Initial commit with working implementation
A python package, backed by Rust for verifiying audio datatsets. While
currently only Wave files are supported, this will hopefully change in
future.

Idea is to verify Wave files by only reading the header and not actually
parsing the wave data itself. Can verify anything contained with the FMT
chunk of a Wave file and the duration of the signal.

Can be run in parallel mode (default) or not by passing
``parallel=False`` in the ``verify_audio`` function.